fix(bundler): dmg volume icon
