+++ .changes/fix-dmg-volume-icon.md
@@ -1,0 +1,5 @@
+---
+"tauri-bundler": patch
+---
+
+Fixes DMG volume icon.
