+++ tooling/bundler/src/bundle/macos/mod.rs
@@ -1,4 +1,5 @@
 pub mod app;
 pub mod dmg;
+pub mod icon;
 pub mod ios;
 pub mod sign;
